// src/resolvers/types/state.ts
export enum State {
  // admin
  chooseConference = 'choose_conference',
  inputCeoPhone = 'input_ceo_phone',

  // ceo
  chooseClub = 'choose_club',
  inputDirectorPhone = 'input_director_phone',

  // director
  inputClubName = 'input_club_name',
  inputClubCity = 'input_club_city',
  inputNewClubName = 'input_new_club_name',
  inputNewClubCity = 'input_new_club_city',
  inputMemberPhone = 'input_member_phone',
  chooseMemberForDelete = 'choose_member_for_delete',
  inputSetName = 'input_set_name',
  inputSetGoal = 'input_set_goal',
  inputSetFinish = 'input_set_finish',
  chooseFriend = 'choose_friend',
  inputNotifyMessage = 'input_notify_message',

  // member
  inputSelfGoal = 'input_self_goal',
}